Img Loads: narrow types, limited support for Extras loading images
PalMod-side ImgDat loading has been upscaling variable width for a long time.  This realigns width to the actual internal ImgDat implementation, as well as the light but extensive follow-up to get the rest of the code on the same page.

Additionally, we now allow Extras files to specify an image unit id and image preview id for most games.  This isn't currently supported if the game has a custom preview loader (UpdatePalImg override) because I didn't want to realign those implementations as well alongside this already broad change.   This new support is probably fairly useless, but wiring it through as part of improving our relationship with Extra files.

+++ palmod/Game/Game_CVS2_A.cpp
@@ -7,7 +7,7 @@
 struct sCVS2_ExtraPair
 {
     LPCWSTR pszExtraName = nullptr;
-    uint16_t nImgIndex = 0;
+    uint8_t nImgIndex = 0;
 };
 
 struct sCVS2_CharacterDump
